tests/wpt/tests/js-self-profiling/js-profiling-mode/unknown.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    test(t => {
      assert_throws_dom('NotAllowedError', () => {
        new Profiler({ sampleInterval: 10, maxBufferSize: Number.MAX_SAFE_INTEGER });
      });
    }, 'js-profiling-mode=unknown (invalid token) should not allow profiling');
  </script>
</body>
</html>
